pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastx-cli"
dynamic = ["version"]
description = "FastMVC CLI — interactive FastAPI project generator and tooling"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Shiv" }]
keywords = ["fastapi", "mvc", "cli", "generator", "fastmvc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "rich>=13.0",
    "fastx-platform>=1.5.0",
    "fastx-middleware>=1.5.0",
    "fastx-database>=1.5.0",
    "fastx-dashboards>=1.5.0",
    "fastx-mvc>=1.5.0",
]

[project.optional-dependencies]
interactive = ["questionary>=2.0"]
dev = [
    "pre-commit>=3.0",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "questionary>=2.0",
    "build>=1.2",
    "twine>=5.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0",
]
all = [
    "fastmvc-cli[interactive,dev]",
    "fastx-platform[all]",
    "fastx-middleware",
    "fastx-database",
    "fastx-dashboards",
    "fastx-mvc[all,dev,docs,interactive]",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=fastx_cli --cov-report=term-missing --cov-fail-under=99"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["fastx_cli"]
branch = false

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["fastx_cli", "tests"]
exclude = ["fastx_cli/bundled"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008", "E501", "B904"]

[tool.mypy]
python_version = "3.10"
packages = ["fastx_cli"]
ignore_missing_imports = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "fastx_cli.validators"
ignore_errors = true

[[tool.mypy.overrides]]
module = "fastx_cli.bundled.git_log_recorder"
ignore_errors = true

[project.urls]
Homepage = "https://github.com/shregar1/fastMVC"
Documentation = "https://github.com/shregar1/fastMVC/blob/main/fastx_cli/README.md"
Repository = "https://github.com/shregar1/fastMVC"
Changelog = "https://github.com/shregar1/fastMVC/blob/main/fastx_cli/CHANGELOG.md"

[project.scripts]
fastx = "fastx_cli.app:main"
fast = "fastx_cli.app:main"
fastx-cli = "fastx_cli.app:main"
fastmvc = "fastx_cli.app:main"

[tool.hatch.version]
path = "fastx_cli/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["fastx_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/pyproject.toml",
    "/fastx_cli",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
    "/CONTRIBUTING.md",
    "/RELEASING.md",
    "/SECURITY.md",
    "/docs",
    "/examples",
    "/.devcontainer",
    "/requirements.txt",
]
